BugFix: when we create a new Conv node in combo's exchange_leader(), this irn has no node_t.
When we create a new Conv node in combo's exchange_leader(), this irn has no node_t.
This is ok, replace the original node in a post walker.
However, as the new node has a visited count of 0 it might be revisited
again.
Fix this by copying the visited count.

This fixes testsuite/opt/fehler179.c

[r28291]